Tried to clean up one-line file state in dbglfs.py
Before:

  littlefs v2.0 0x{0,1}.232, rev 99, weight 9.256, bd 4096x256
  {00a3,00a4}:   0.1 file0000  reg 32768, trunk 0xa3.a8 32768, btree 0x1a.846 32704
                 0.2 file0001  reg 32768, trunk 0xa3.16c 32768, btree 0xa2.be1 32704

After:

  littlefs v2.0 0x{0,1}.232, rev 99, weight 9.256, bd 4096x256
  {00a3,00a4}:   0.1 file0000  reg 32768, trunk 0xa3.a8, btree 0x1a.846
                 0.2 file0001  reg 32768, trunk 0xa3.16c, btree 0xa2.be1

Most files will have both a shrub and a btree, which makes the previous
output problematically noisy.

Unfortunately, this does lose some information: the size of the
shrub/tree, both of which may be less than the full file. But 1. this
is _technically_ redundant since you only need the block/trunk to fetch an
rbyd (though the weight is useful), and 2. The weight can still be
viewed with -s -i.