Implemented a block crystalization algorithm that actually works
The idea here is to:

1. Try to figure out the current "crystal" (set of fragments) we are a
   part of.

2. Decide if our crystal has probably exceeded the configured crystal size
   and needs to be compacted into a block.

3. Guess the local block alignment by looking at the entry immediately
   left of our crystal.

Figuring out the state of our current crystal is done heuristically,
with a lookup 1 crystal-size to the left to find the start of our
crystal, followed by a lookup 1 crystal-size to the right of the crystal
start to find the end of our crystal:

                     -crs    pos        -crs    pos
                     .-------|          .-------|
      -crs    pos    '---.   |   +crs   '-------.      +crs
      .-------|          |---|---.              |------.
  .---'       |          |   |   '---.          |  .---'
  v           v          v   v       v          v  v
  .---+---+---+---.  +---.---+---+---.  ---+---.---.---+
  | crystal       |  blk | crystal   |  blk    |crs| blk
  '---+---+---+---'  +---'---+---+---'  ---+---'---'---+

This is a heuristic that doesn't catch any holes in our crystal, but
that's ok, we probably don't want small holes preventing block
compaction anyways.

Finding the block alignment then just requires looking up the entry to
the left of our crystal, if the left entry + crystal fits in a block, we
are the same block, otherwise we align to the left entry. Note this may
break our crystal during block compaction if the crystal itself is not
block aligned, but that's ok, we just recalculate the new crystal based
on the new block:

         fits             partial fit            doesn't fit
  .---+---+---+---.  .---+---+---+---+---.  .---+---+---+---+---.
  | block     |crs|  | block     | crs   |  | block         |crs|
  '---+---+---+---'  '---+---+---+---+---'  '---+---+---+---+---'
          |                    |                      |
          v                    v                      v
  .---+---+---+---.  .---+---+---+---+---.  .---+---+---+---+---+- - -
  | block         |  | block         |crs|  | block         | block
  '---+---+---+---'  '---+---+---+---+---'  '---+---+---+---+---+- - -

This involves at most 3 lookups, though there are some shortcuts: If
appending a file we never need to lookup the right crystal boundary,
and if we don't exceed our crystal size we don't need to figure out the
block alignment.

---

There's another variant of this scheme where we don't consider any
fragments to the right of the current fragment. This saves a lookup, but
more importantly would mean we could take advantage of ecksums when
partially rewriting part of a file.

As a tradeoff this variant does end up leaving any partially rewritten
files with >~2x storage overhead. Still, this may be interesting to
provide as an alternative write strategy in the future.