Reverted moving lfsr_mdir_drop out of lfsr_mdir_commit
The logic is simpler with lfsr_mdir_drop being an implicit side-effect
of lfsr_mdir_commit, and right now a simpler system is preferred over a
complex one.

I think it will be more useful to focus on reducing the stack overhead
of lfsr_mdir_drop in general.

Still, it may be worth reverting this revert in the future.