Moved post-compaction mdir commits into lfsr_mdir_compact__
Since we need access to the pending attr-list in lfsr_mdir_compact__
now, we might as well just do the pending commit. Worst case, a call to
lfsr_mdir_compact__ can provide a NULL attr-list for the previous
behavior (though we always follow up compaction with a commit).

The only downside is that this behavior is now a bit different from
lfsr_rbyd_compact. Need to revisit lfsr_rbyd_compact and see if it
should do the same.

This had a tiny improvement to code size, which I think is just the cost
of two function calls:

            code          stack
  before:  31716           2776
  after:   31676 (-0.1%)   2776 (+0.0%)